pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepseek-ocr-2"
version = "0.1.0"
description = "Mechanistic interpretability toolkit for DeepSeek-OCR-2"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.6.0",
    "torchvision",
    "transformers==4.46.3",
    "tokenizers==0.20.3",
    "einops",
    "addict",
    "easydict",
    "pillow",
    "numpy",
    "safetensors",
    "huggingface-hub",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "imageio>=2.31.0",
    "pytest>=9.0.2",
    "jupyter>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "jupyter",
    "ipykernel",
]
animation = [
    "imageio>=2.31.0",
]
